Fix unit test with conftest
Apparently, fixture with autouse=True is not called when the fixture is in conftest.py
