+++ backend/tests/test_api.py
@@ -1,11 +1,22 @@
 import json
 from decimal import Decimal
+import os
 from unittest.mock import patch
 
 import boto3
 import pytest
 from fastapi.testclient import TestClient
 from moto import mock_aws
+
+
+@pytest.fixture
+def aws_credentials(scope="session", autouse=True):
+    """Mocked AWS Credentials for moto."""
+    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
+    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
+    os.environ["AWS_SECURITY_TOKEN"] = "testing"
+    os.environ["AWS_SESSION_TOKEN"] = "testing"
+    os.environ["LOG_LEVEL"] = "TRACE"
 
 
 @pytest.fixture
