+++ backend/tests/test_upload.py
@@ -1,8 +1,19 @@
 import json
+import os
 
 import boto3
 import pytest
 from moto import mock_aws
+
+
+@pytest.fixture
+def aws_credentials(scope="session", autouse=True):
+    """Mocked AWS Credentials for moto."""
+    os.environ["AWS_ACCESS_KEY_ID"] = "testing"
+    os.environ["AWS_SECRET_ACCESS_KEY"] = "testing"
+    os.environ["AWS_SECURITY_TOKEN"] = "testing"
+    os.environ["AWS_SESSION_TOKEN"] = "testing"
+    os.environ["LOG_LEVEL"] = "TRACE"
 
 
 @mock_aws
